Reorganisation of files
Added folder for common files such as constants, utils, apis, ...
